[rpc] Add WebSocket metrics
